Refactor galaxy.tools.provided_metadata -> galaxy.tool_util.
Needed for remote job output metadata collection.
