+++ lib/galaxy/tool_util/provided_metadata.py
